Fix bug caused by page visibility
